services/chat/requirements.txt
# Chat Service Dependencies
fastapi==0.115.5
uvicorn[standard]==0.30.0
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.27.0
asyncpg==0.30.0
redis[hiredis]==5.2.1
PyJWT[crypto]==2.9.0
structlog==24.4.0

